+++ auto_derby/texttools.py
@@ -133,15 +133,15 @@ def compare(a: Text, b: Text, *, max_missing_chars: int = 5) -> float:
 
 
 def choose(v: Text, options: Iterable[Text], threshold: float = 0.5) -> Text:
-    option_with_similarites = sorted(
+    option_with_similarities = sorted(
         ((i, compare(v, i)) for i in options),
         key=lambda x: x[1],
         reverse=True,
     )
-    if not option_with_similarites:
+    if not option_with_similarities:
         app.log.text("choose: empty options", level=app.WARN)
         return v
-    res, similarity = option_with_similarites[0]
+    res, similarity = option_with_similarities[0]
     app.log.text(
         "choose: text='%s' match='%s' similarity=%.3f threshold=%.3f"
         % (